// src/lib/types/productType.ts
export type ProductType = {
  productID: string
  productName: string
  mediaURL: string
  description: string
  createdDate: string
  discountActive: boolean
  discountEndDate: string
  discountStartDate: string
  discountValuePercent: number
  minimumOrderQuantity: number
  condition: string
  finalPrice: number
  price: number
}